Refactor and enhance using OpensrfClient, Response
Refactor to move from the OpensrfClientRequest object to
an OpensrfClient object:
* has an endpoint (the gateway)
* caches a curl resource
* request method takes service, method, params

OpensrfResponse object
* returned by the OpensrfClient->request() method
* all properties for now:
 - success
 - http_status
 - raw
 - status
 - payload
 - debug

Also enhanced error handling a bit to notice http-level errors.

If there's a curl error (ssl error, can't connect to host),
the response object's success property is false and we get
a curl error in the debug property.

If there's an HTTP error (404 due to bad service/method, etc),
the response object's success property is false and we get an
http error in the debug property. We also set the http_status
property.

As before, if there's an OpenSRF error, we get the status code
in the status property, and the debug property usually contains
a helpful application error.